+++ changelog.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add delete shortcut to remove components or links.
 - Add read-only option to draw function.
 - Add resizing on draw function when read-only is activated.
+- Add `tags` attribute in `DefaultConfiguration`.
 
 ## [0.15.2] - 2023/04/18